+++ WadTool/FormMain.cs
@@ -262,7 +262,7 @@ namespace WadTool
 
             if (_playAnimation)
             {
-                int frameRate = (int)Math.Max(animation.FrameDuration, (short)1);
+                int frameRate = Math.Max(animation.FrameDuration, (short)1);
                 timerPlayAnimation.Stop();
                 timerPlayAnimation.Interval = 1000 / 30 / frameRate;
                 timerPlayAnimation.Start();
@@ -448,7 +448,7 @@ namespace WadTool
 
             // Start animation
             _playAnimation = true;
-            int frameRate = (int)Math.Max(animation.FrameDuration, (short)1);
+            int frameRate = Math.Max(animation.FrameDuration, (short)1);
             timerPlayAnimation.Stop();
             timerPlayAnimation.Interval = 1000 / 30 / frameRate;
             timerPlayAnimation.Start();
